# docs/output/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: 예시 문서입니다.
  description: 예시 API
  title: Example API
  version: 1.0.0
servers:
- description: 개발용 로컬 서버
  url: http://localhost:8080
- description: 개발용 목 서버
  url: http://localhost:8081
...
paths:
...
  /api/v1/projects:
    get:
      operationId: getAllProjects
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrudProjectPageDto'
          description: 프로젝트 목록
        "404":
          content:
            application/json:
              example:
                type: http://localhost:8080/api/v1
                title: 요청한 리소스를 찾을 수 없습니다.
                status: 404
                detail: ID가 '12345'인 API를 데이터베이스에서 찾을 수 없습니다.
                instance: projects/2/apis/12345
                properties:
                  code: API4001
                  isSuccess: "false"
              schema:
                $ref: '#/components/schemas/NotFoundException'
          description: 요청한 리소스를 찾을 수 없습니다.
      summary: 프로젝트 전체 목록 조회
      tags:
      - ScrudProject
      x-spring-paginated: true
    patch:
      description: "프로젝트 제목, 설명, 서버url 수정"
      operationId: updateScrudProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScrudProjectDto'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScrudProjectDto'
          description: 프로젝트 설정
      tags:
      - ScrudProject
    post:
      description: "프로젝트 생성 API. 전역 파일 추가 주의사항! 아키텍처 구조도는 사용자 선택일 경우 ARCHITECTURE_DEFAULT,\
        \ 깃허브 연동해서 올리는 경우엔 ARCHITECTURE_GITHUB 로 부탁해요! 그리고 fileContent: JSON.stringify(fileContent)\
        \ 이렇게 직렬화해서 보내주세요!"
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                format: int64
                type: integer
          description: ProjectId 반환
      summary: 프로젝트 생성
      tags:
      - ScrudProject
  /api/v1/projects/{projectId}:
    delete:
      description: 프로젝트 삭제
      operationId: deleteProject
      parameters:
      - description: 프로젝트 ID
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: 성공적으로 처리되었습니다
      tags:
      - ScrudProject
    get:
      operationId: getAllGlobalFile
      parameters:
      - description: 프로젝트 ID
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalFileListDto'
          description: 전체 전역설정 파일 목록
        "404":
          content:
            application/json:
              example:
                type: http://localhost:8080/api/v1
                title: 요청한 리소스를 찾을 수 없습니다.
                status: 404
                detail: ID가 '12345'인 API를 데이터베이스에서 찾을 수 없습니다.
                instance: projects/2/apis/12345
                properties:
                  code: API4001
                  isSuccess: "false"
              schema:
                $ref: '#/components/schemas/NotFoundException'
          description: 요청한 리소스를 찾을 수 없습니다.
      tags:
      - ScrudProject
    put:
      description: "전역 파일 추가 - 아키텍처 구조도는 사용자 선택일 경우 ARCHITECTURE_DEFAULT, 깃허브 연동해서\
        \ 올리는 경우엔 ARCHITECTURE_GITHUB 로 부탁해요! 그리고 fileContent: JSON.stringify(fileContent)\
        \ 이렇게 직렬화해서 보내주세요!"
      operationId: addGlobalFile
      parameters:
      - description: 프로젝트 ID
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalFileDto'
      responses:
        "204":
          description: 성공적으로 처리되었습니다
      tags:
      - ScrudProject
  /api/v1/projects/{projectId}/{globalFileId}:
    delete:
      description: 전역 파일 삭제
      operationId: deleteGlobalFile
      parameters:
      - description: 프로젝트 ID
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: 전역 파일 ID
        explode: false
        in: path
        name: globalFileId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: 성공적으로 처리되었습니다
      tags:
      - ScrudProject
...
components:
...
  schemas:
...
    NotFoundException:
      allOf:
      - $ref: '#/components/schemas/ProblemDetail'
      - properties:
          code:
            type: string
          isSuccess:
            type: boolean
        type: object
      example:
        instance: https://openapi-generator.tech
        code: code
        detail: detail
        type: https://openapi-generator.tech
        title: title
        status: 6
        isSuccess: true
    ProblemDetail:
      properties:
        type:
          description: 문제 유형을 식별하는 URI
          format: uri
          type: string
        title:
          description: 문제에 대한 간단한 제목
          type: string
        status:
          description: HTTP 상태 코드
          type: integer
        detail:
          description: 문제에 대한 자세한 설명
          type: string
        instance:
          description: 문제가 발생한 특정 URI
          format: uri
          type: string
      type: object
...
    ScrudProjectPageDto:
      description: 프로젝트 목록
      example:
        pageable:
          isLastPage: true
          totalPages: 1
          listSize: 6
          isFirstPage: true
          totalElements: 5
        content:
        - isDeleted: true
          serverUrl: serverUrl
          scrudProjectId: 0
          description: description
          title: title
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - isDeleted: true
          serverUrl: serverUrl
          scrudProjectId: 0
          description: description
          title: title
          updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        content:
          items:
            $ref: '#/components/schemas/ScrudProjectDto'
          type: array
        pageable:
          $ref: '#/components/schemas/PageDto'
      required:
      - content
      - pageable
      type: object
      x-spring-paginated: true
    ScrudProjectDto:
      description: 프로젝트 기본 정보
      example:
        isDeleted: true
        serverUrl: serverUrl
        scrudProjectId: 0
        description: description
        title: title
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        scrudProjectId:
          description: 프로젝트 ID
          format: int64
          type: integer
        title:
          description: 프로젝트 제목
          type: string
        description:
          description: 프로젝트 설명
          type: string
        serverUrl:
          description: serverUrl
          type: string
        updatedAt:
          description: 프로젝트 수정 날짜
          format: date-time
          type: string
        isDeleted:
          description: 프로젝트 삭제 여부
          type: boolean
      type: object
    CreateProjectRequest:
      description: 프로젝트 생성에 필요한 기본 정보 & 전역 설정
      example:
        globalFiles:
        - fileName: fileName
          globalFileId: 0
          fileType: REQUIREMENTS
          fileContent: fileContent
        - fileName: fileName
          globalFileId: 0
          fileType: REQUIREMENTS
          fileContent: fileContent
        scrudProjectDto:
          isDeleted: true
          serverUrl: serverUrl
          scrudProjectId: 0
          description: description
          title: title
          updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        scrudProjectDto:
          $ref: '#/components/schemas/ScrudProjectDto'
        globalFiles:
          items:
            $ref: '#/components/schemas/GlobalFileDto'
          type: array
      type: object
    GlobalFileDto:
      description: 전역설정 파일 목록
      example:
        fileName: fileName
        globalFileId: 0
        fileType: REQUIREMENTS
        fileContent: fileContent
      properties:
        globalFileId:
          format: int64
          type: integer
        fileName:
          type: string
        fileType:
          $ref: '#/components/schemas/FileTypeEnumDto'
        fileContent:
          type: string
      type: object
    FileTypeEnumDto:
      enum:
      - REQUIREMENTS
      - ERD
      - UTIL
      - CONVENTION
      - CONVENTION_DEFAULT
      - DEPENDENCY
      - ERROR_CODE
      - SECURITY
      - SECURITY_DEFAULT_JWT
      - SECURITY_DEFAULT_SESSION
      - SECURITY_DEFAULT_NONE
      - ARCHITECTURE_GITHUB
      - ARCHITECTURE_DEFAULT_LAYERED_A
      - ARCHITECTURE_DEFAULT_LAYERED_B
      - ARCHITECTURE_DEFAULT_CLEAN
      - ARCHITECTURE_DEFAULT_MSA
      - ARCHITECTURE_DEFAULT_HEX
      type: string
    GlobalFileListDto:
      description: 전체 전역설정 파일 목록
      example:
        project:
          isDeleted: true
          serverUrl: serverUrl
          scrudProjectId: 0
          description: description
          title: title
          updatedAt: 2000-01-23T04:56:07.000+00:00
        content:
        - fileName: fileName
          globalFileId: 0
          fileType: REQUIREMENTS
          fileContent: fileContent
        - fileName: fileName
          globalFileId: 0
          fileType: REQUIREMENTS
          fileContent: fileContent
      properties:
        project:
          $ref: '#/components/schemas/ScrudProjectDto'
        content:
          items:
            $ref: '#/components/schemas/GlobalFileDto'
          type: array
      type: object
...
    PageDto:
      example:
        isLastPage: true
        totalPages: 1
        listSize: 6
        isFirstPage: true
        totalElements: 5
      properties:
        listSize:
          description: 현재 페이지의 데이터 개수
          type: integer
        isFirstPage:
          description: 현재 페이지가 첫 번째 페이지인지 여부
          type: boolean
        isLastPage:
          description: 현재 페이지가 마지막 페이지인지 여부
          type: boolean
        totalPages:
          description: 총 페이지 수
          type: integer
        totalElements:
          description: 총 데이터 개수
          type: integer
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: "JWT 인증 토큰을 사용합니다. 예: Bearer {token}"
      scheme: bearer
      type: http